+++ GLPM.TechnicalAnalysis/TA/TAFunc/TA_Aroon.cs
@@ -67,7 +67,7 @@ namespace GLPM.TechnicalAnalysis
             int highestIdx = -1;
             double lowest = 0.0;
             double highest = 0.0;
-            double factor = 100.0 / ((double)optInTimePeriod);
+            double factor = 100.0 / optInTimePeriod;
             Label_00BB:
             if (today > endIdx)
             {
@@ -209,7 +209,7 @@ namespace GLPM.TechnicalAnalysis
             int highestIdx = -1;
             double lowest = 0.0;
             double highest = 0.0;
-            double factor = 100.0 / ((double)optInTimePeriod);
+            double factor = 100.0 / optInTimePeriod;
             Label_00BB:
             if (today > endIdx)
             {
