+++ GLPM.TechnicalAnalysis/TA/TAFunc/TA_Cosh.cs
@@ -78,7 +78,7 @@ namespace GLPM.TechnicalAnalysis
             int outIdx = 0;
             while (i <= endIdx)
             {
-                outReal[outIdx] = Math.Cosh((double)inReal[i]);
+                outReal[outIdx] = Math.Cosh(inReal[i]);
                 i++;
                 outIdx++;
             }
